pushing Footer Component (reorganised) ---by Pawan
