+++ sugar-cosmetics/src/App.js
@@ -1,4 +1,4 @@
-import { Footer } from "./Components/Footer";
+import { Footer } from "./Components/Footer/Footer";
 import "./styles.css";
 
 export default function App() {
